requirements.txt
Django>=4.0.5
Faker>=13.15.0
django-filter>=22.1